% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votings.R
\name{clear_cache}
\alias{clear_cache}
\title{Clears the cache directory.}
\usage{
clear_cache(data_dir = "~/.riigikogu/votings/")
}
\arguments{
\item{data_dir}{a cache directory that holds the downloaded data files.}
}
\description{
Clears the cache directory.
}
\examples{
clear_cache()
}
