% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votings.R
\name{get_votes}
\alias{get_votes}
\title{Get all members' votes for a specific voting.}
\usage{
get_votes(uuid, data_dir = "~/.riigikogu/votings/")
}
\arguments{
\item{uuid}{id of the voting.}

\item{data_dir}{a cache directory for the downloaded data files.}
}
\value{
A dataframe listing information about the votings within the time period.
}
\description{
Get all members' votes for a specific voting.
}
\examples{
uuid = "1a021b93-c629-4c2b-99ce-15f77596bfc3"
votes = get_votes(uuid)
}
