% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votings.R
\name{get_votings}
\alias{get_votings}
\title{Get all votings within a given date range.}
\usage{
get_votings(start_date, end_date)
}
\arguments{
\item{start_date}{start of the date range (yyyy-MM-dd).}

\item{end_date}{end of the date range (yyyy-MM-dd).}
}
\value{
A dataframe listing information about votings within the time period.
}
\description{
Get all votings within a given date range.
}
\examples{
start_date = "2016-11-23"
end_date = "2019-04-29"
votings = get_votings(start_date, end_date)
}
