% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votings.R
\name{is_in_coalition}
\alias{is_in_coalition}
\title{Determines whether a parliament member was in coaliation between a given date range.}
\usage{
is_in_coalition(nodes, start_date, end_date)
}
\arguments{
\item{nodes}{a dataframe containing summary information about parliament members.}

\item{start_date}{start of the date range (yyyy-MM-dd).}

\item{end_date}{end of the date range (yyyy-MM-dd).}
}
\value{
A boolean vector indicating whether the member was in coaliation.
}
\description{
Determines whether a parliament member was in coaliation between a given date range.
}
\examples{
start_date = "2016-11-23"
end_date = "2019-04-29"

votings = get_votings(start_date, end_date)
votings = filter(votings, type.code == "AVALIK")

votes_list = lapply(votings$uuid, get_votes)
has_no_votes = sapply(votes_list, is.null)
votes_list = votes_list[!has_no_votes]

nodes = get_nodes(votes_list)
nodes$in_coalition = is_in_coalition(nodes, start_date, end_date)
}
